Add initial profiler classes
This change makes the profiler class an interface. The reason to add
the interface is to allow for polymorphic dispatch of one or many
profilers to collect metrics from many parts of the application stack.

The workload profiler will eventually support all application types
on EMR.

Pull requests are always welcome!