Add initial set of collectors
This change adds the initial set of profiler classes
and changes the profiler class to an interface.

The profiler will support collecting from all of the different
application types that run on EMR. The interface is created
to make it simple and easy to add new profilers when
new application types are supported on EMR.

Pull requests are always welcome.