Add main entry point
This change adds the main entry point for the application and introduces
the following: logging, configuration file parsing, and creation
and execution of profilers.

This change also adds the shade plug-in to the pom file in order to
build a jar with all dependencies included. This can be built with the
mvn package command and executed with java -jar jarname.jar.